Move processes tooltip and replace the content
